requirements.txt
pytest
pylint
ipdb
ipykernel
yapf
matplotlib
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
requests
google-cloud-bigquery
google-cloud-storage
db-dtypes
pyarrow
